# public/apidoc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APD-API",
        "description": "An API to perform air pressure drop calculation through projects and user account",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://aeraulic.io/api/V1",
            "description": "Apd calculator API server"
        }
    ],
    "paths": {
...
        "/apd/ductnetworks/{id}/ductsections": {
            "get": {
                "tags": [
                    "Duct section"
                ],
                "summary": "Get all duct sections",
                "description": "Get all duct sections from a duct network",
                "operationId": "e22e325b085af24814b62cc80ee68478",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to get all duct sections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ductSection"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/apd/ductnetworks/{ductNetworkId}/ductsections/{ductSectionId}": {
            "get": {
                "tags": [
                    "Duct section"
                ],
                "summary": "Get one duct section",
                "description": "Get one duct section with his id from a duct network",
                "operationId": "3fb1e7dd42600591277c80545b5dfad9",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ductNetworkId"
                    },
                    {
                        "$ref": "#/components/parameters/ductSectionId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to get one duct section",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/ductSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Duct section"
                ],
                "summary": "Remove a duct section",
                "description": "Remove a duct section in a duct network",
                "operationId": "a7a72b4e85a119b415f6f7947804ad03",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ductNetworkId"
                    },
                    {
                        "$ref": "#/components/parameters/ductSectionId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to remove a duct section",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "DuctSection with name \"section DE\" has been deleted."
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/ductSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Duct section"
                ],
                "summary": "Update a duct section",
                "description": "Update a duct section in a duct network",
                "operationId": "ad2829cb4ff3a06a8d8294ba1f7d983f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ductNetworkId"
                    },
                    {
                        "$ref": "#/components/parameters/ductSectionId"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateDuctSection"
                },
                "responses": {
                    "200": {
                        "description": "Success to update a duct section",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/ductSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessableEntity"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/apd/ductnetworks/{ductNetworkId}/ductsections": {
            "post": {
                "tags": [
                    "Duct section"
                ],
                "summary": "Add a duct section",
                "description": "Add a duct section in a duct network",
                "operationId": "2ebc23fe69e34531059b8315eeb9c0f7",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ductNetworkId"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/addDuctSection"
                },
                "responses": {
                    "200": {
                        "description": "Success to add a duct section",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/ductSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessableEntity"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ductSection": {
                "title": "duct section",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "duct section id",
                        "type": "integer",
                        "example": 84
                    },
                    "name": {
                        "title": "name",
                        "description": "Duct network's name",
                        "type": "string",
                        "example": "section DE"
                    },
                    "ductNetworkId": {
                        "title": "duct network id",
                        "description": "Duct section's associated duct network id",
                        "type": "integer",
                        "example": 42
                    },
                    "air": {
                        "title": "air",
                        "description": "Air instance of the duct network",
                        "properties": {
                            "viscosity": {
                                "title": "viscosity",
                                "description": "Viscosity property of the air",
                                "type": "number",
                                "format": "float",
                                "example": 1.5080510051843115e-5
                            },
                            "density": {
                                "title": "density",
                                "description": "Density property of the air",
                                "type": "number",
                                "format": "float",
                                "example": 1.2058928673556562
                            },
                            "temperature": {
                                "title": "temperature",
                                "description": "Temperature property of the air",
                                "type": "number",
                                "format": "float",
                                "example": 18.2
                            },
                            "altitude": {
                                "title": "altitude",
                                "description": "Altitude property of the air",
                                "type": "integer",
                                "example": 800
                            }
                        },
                        "type": "object"
                    },
                    "shape": {
                        "title": "shape",
                        "description": "Duct desction's shape - circular or rectangular",
                        "type": "string",
                        "pattern": "^circular|rectangular$",
                        "example": "circular"
                    },
                    "material": {
                        "title": "material",
                        "description": "Duct desction's material",
                        "type": "string",
                        "example": "galvanised_steel"
                    },
                    "flowrate": {
                        "title": "flow rate",
                        "description": "Duct section's flow rate _ in cubic meter per hour (m³/h)",
                        "type": "integer",
                        "minimum": 100,
                        "example": 2500
                    },
                    "length": {
                        "title": "length",
                        "description": "Duct section's linear length - in meter (m)",
                        "type": "number",
                        "format": "float",
                        "minimum": 0.1,
                        "example": 5.4
                    },
                    "singularities": {
                        "title": "singularities",
                        "description": "List and number of singularities in the duct section",
                        "type": "array",
                        "items": {
                            "properties": {
                                "singularity": {
                                    "title": "singularity type",
                                    "type": "string",
                                    "example": "90_elbow"
                                },
                                "number": {
                                    "title": "number of singularity",
                                    "type": "integer",
                                    "example": 2
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "90_elbow": 2
                        }
                    },
                    "additionalApd": {
                        "title": "additional apd",
                        "description": "Optional additional air pressure drop value who represent an accessory of duct section - in pascal (Pa)",
                        "type": "integer",
                        "minimum": 1,
                        "example": 50
                    },
                    "diameter": {
                        "title": "diameter",
                        "description": "Duct section's diameter when shape is circular (normalized diameter) - in millimeter (mm)",
                        "type": "integer",
                        "example": 315
                    },
                    "width": {
                        "title": "width",
                        "description": "Duct section's width when shape is rectangular - in millimeter (mm)",
                        "type": "integer",
                        "example": 400
                    },
                    "height": {
                        "title": "height",
                        "description": "Duct section's height when shape is rectangular - in millimeter (mm)",
                        "type": "integer",
                        "example": 300
                    },
                    "equivDiameter": {
                        "title": "equiv diameter",
                        "description": "Duct section equivalent diameter calculation - in millimeter (mm)",
                        "type": "number",
                        "format": "float",
                        "example": 380.697
                    },
                    "ductSectionsSection": {
                        "title": "duct section's section",
                        "description": "Section of the duct section calculation - in square meter (m²)",
                        "type": "number",
                        "format": "float",
                        "example": 0.114
                    },
                    "flowspeed": {
                        "title": "flow speed",
                        "description": "Duct section's flow speed calculation - in meter per second (m/s)",
                        "type": "number",
                        "format": "float",
                        "example": 7.31
                    },
                    "linearApd": {
                        "title": "linear apd",
                        "description": "Duct section's linear apd calculation - in pascal (Pa)",
                        "type": "number",
                        "format": "float",
                        "example": 11.011
                    },
                    "singularApd": {
                        "title": "singular apd",
                        "description": "Duct section's singular apd calculation - in pascal (Pa)",
                        "type": "number",
                        "format": "float",
                        "example": 9.693
                    },
                    "totalApd": {
                        "title": "total apd",
                        "description": "Duct section's total apd calculation - in pascal (Pa)",
                        "type": "number",
                        "format": "float",
                        "example": 70.704
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "notFound": {
                "description": "Ressource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Error"
                                },
                                "content": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "id"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Ressource doesn't exist with this id."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Ressource's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "ductNetworkId": {
                "name": "ductNetworkId",
                "in": "path",
                "description": "Duct network's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "ductSectionId": {
                "name": "ductSectionId",
                "in": "path",
                "description": "Duct section's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
...
            "updateDuctSection": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "title": "name",
                                    "description": "Name of the duct network",
                                    "type": "string",
                                    "example": "duct network n°1",
                                    "nullable": true
                                },
                                "material": {
                                    "title": "material",
                                    "description": "Duct desction's material",
                                    "type": "string",
                                    "example": "galvanised_steel",
                                    "nullable": true,
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/materialsName"
                                        }
                                    ]
                                },
                                "shape": {
                                    "title": "shape",
                                    "description": "Duct desction's shape - circular or rectangular",
                                    "type": "string",
                                    "pattern": "^circular|rectangular$",
                                    "example": "circular",
                                    "nullable": true
                                },
                                "flowrate": {
                                    "title": "flow rate",
                                    "description": "Duct section's flow rate _ in cubic meter per hour (m³/h)",
                                    "type": "integer",
                                    "minimum": 100,
                                    "example": 2500,
                                    "nullable": true
                                },
                                "length": {
                                    "title": "length",
                                    "description": "Duct section's linear length - in meter (m)",
                                    "type": "number",
                                    "format": "float",
                                    "minimum": 0.1,
                                    "example": 5.4,
                                    "nullable": true
                                },
                                "singularities": {
                                    "title": "singularities",
                                    "description": "List and amount of singularities in the duct section",
                                    "type": "object",
                                    "nullable": true,
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/singularitiesAmount"
                                        }
                                    ]
                                },
                                "additionalApd": {
                                    "title": "additional apd",
                                    "description": "Optional additional air pressure drop value who represent an accessory of duct section - in pascal (Pa)",
                                    "type": "integer",
                                    "minimum": 1,
                                    "example": 50,
                                    "nullable": true
                                },
                                "diameter": {
                                    "title": "diameter",
                                    "description": "Duct section's diameter when shape is circular (normalized diameter) - in millimeter (mm)",
                                    "type": "integer",
                                    "example": 315,
                                    "nullable": true,
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/diameters"
                                        }
                                    ]
                                },
                                "width": {
                                    "title": "width",
                                    "description": "Duct section's width when shape is rectangular - in millimeter (mm)",
                                    "type": "integer",
                                    "example": 400,
                                    "nullable": true
                                },
                                "height": {
                                    "title": "height",
                                    "description": "Duct section's height when shape is rectangular - in millimeter (mm)",
                                    "type": "integer",
                                    "example": 300,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}